requirements.txt
Django==1.8.1
django-filter==0.10.0
djangorestframework==3.1.2
Markdown==2.6.2
(wcp)Edwards-Air